debugging

+++ ft_printf/srcs/helper2.c
@@ -6,11 +6,12 @@
 /*   By: hroxo <hroxo@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/11 13:37:07 by hroxo             #+#    #+#             */
-/*   Updated: 2025/09/13 23:52:44 by hroxo            ###   ########.fr       */
+/*   Updated: 2025/09/14 11:30:02 by hroxo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stddef.h>
+#include <stdlib.h>
 #include "../util/header.h"
 
 size_t	ft_putunsigned_int(unsigned int nb)
@@ -31,7 +32,7 @@ size_t	ft_putpointer(void *nb)
 	size_t	count;
 
 	count = ft_putstr("0x");
-	count += ft_putnbr_base((int)(long)nb, "0123456789abcdef");
+	count += ft_putnbr_base((long)nb, "0123456789abcdef");
 	return (count);
 }
